Move most of the code from DetailViewController so that it can be shared with other VCs.
The logic for synching a list to Swim moves to SwimListManager.  This
knows how to serialize and deserialize model objects as ReconValues,
and manages the synching and the update / move / delete / highlight
messages on the wire.

The logic for taking a Swim list and showing it as a UITableViewController
moves to SwimListTableViewController, which therefore becomes a base class
for DetailViewController (and any other similar VC).

This also adds Logging.swift, which is a very basic debug logging module.
We need a better plan for logging.