Added SwimListViewController.  This is a base class that's very similar to SwimListTableViewController, though without inheriting from UITableViewController.
Use this to take the ClearStyle tutorial code and merge it into
DetailViewController so that we get the ClearStyle gestures and cell
handling for edits and deletes.  Some tweaks have been made along the way.

This doesn't entirely work (inserts are buggy, deletes during a cell edit
will cause an assert) but it's getting there.