Add a dirty hack to the handling for insert messages.
Since we have no session ID or device ID in the on-wire message, we can't
identify our own messages from other people's, which means that we can't
tell whether a new insert is ours (and therefore already performed on the
client side) or someone else's.  This means that we end up inserting the
item twice.

Work around this by simply not inserting on the client side for now.  This
causes a nasty UI glitch when the table is reloaded (removing the cell)
and then the insert arrives on the wire (re-inserting the cell) but it
will have to do for now.

This needs fixing properly.